Add game over communications
First, there was added snake collisions with itself detection. Next,
if the snake hits the edge of the screen or itself, display crash
communication. If the snake has collected wrong brick, display wrong PI
number communication. When the communication is displayed, snake
movement is stopped. Hence it won't change for the other communication.